refactor: domain, infra 패키지 분리
