+++ src/main/java/kitchenpos/domain/menu/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
