+++ src/main/java/kitchenpos/infra/blackword/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.infra.blackword;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
